CinematicTrailerGenAI/requirements.txt
python-dotenv
fal-client
requests
moviepy>=2
imageio-ffmpeg>=0.4.9
